Unify IppRequest and IppResponse
